# Refactor validator