- Removes mocha.
- Switches to async (parallel) validation.
- Counts services instead of documents.
- Makes selector failure detection more accurate.
- Removes “pending tests” workaround.
- 4x speedup.

## Before

### Mean time: 3 minutes 34 seconds

Run | Passing | Pending | Failing | Real | User | Sys
----|---------|---------|---------|------|------|----
1   | 572     | 10      | 5       | 3m51.348s | 1m11.918s | 0m4.362s
2   | 580     |  4      | 3       | 3m14.648s | 1m5.369s | 0m4.276s
3   | 580     |  4      | 3       | 3m37.634s | 1m13.709s | 0m4.470s

![Capture d’écran 2020-08-31 à 10 44 39](https://user-images.githubusercontent.com/222463/91704730-4be51d80-eb7c-11ea-90ff-880692f904d4.png)
![Capture d’écran 2020-08-31 à 10 44 47](https://user-images.githubusercontent.com/222463/91704739-4d164a80-eb7c-11ea-9bf3-4974b50909f8.png)
![Capture d’écran 2020-08-31 à 10 44 55](https://user-images.githubusercontent.com/222463/91704741-4daee100-eb7c-11ea-9fb1-62f408917cb6.png)
![Capture d’écran 2020-08-31 à 10 45 02](https://user-images.githubusercontent.com/222463/91704744-4e477780-eb7c-11ea-9777-8782ee1887c1.png)
![Capture d’écran 2020-08-31 à 10 45 12](https://user-images.githubusercontent.com/222463/91704749-4ee00e00-eb7c-11ea-9e72-93e4a1dbf8f5.png)
![Capture d’écran 2020-08-31 à 10 45 20](https://user-images.githubusercontent.com/222463/91704751-4ee00e00-eb7c-11ea-8aab-b1e8a2e7e529.png)
![Capture d’écran 2020-08-31 à 10 45 26](https://user-images.githubusercontent.com/222463/91704754-4f78a480-eb7c-11ea-97c9-f8925aa936fc.png)
![Capture d’écran 2020-08-31 à 10 45 33](https://user-images.githubusercontent.com/222463/91704756-4f78a480-eb7c-11ea-837d-1ddcdfd13291.png)
![Capture d’écran 2020-08-31 à 10 45 44](https://user-images.githubusercontent.com/222463/91704757-50113b00-eb7c-11ea-9452-d55433773807.png)
![Capture d’écran 2020-08-31 à 10 45 50](https://user-images.githubusercontent.com/222463/91704758-50113b00-eb7c-11ea-92f4-91746334bab0.png)
![Capture d’écran 2020-08-31 à 10 46 00](https://user-images.githubusercontent.com/222463/91704759-50113b00-eb7c-11ea-9310-b445e1e49c27.png)
![Capture d’écran 2020-08-31 à 10 46 08](https://user-images.githubusercontent.com/222463/91704761-50a9d180-eb7c-11ea-84bd-1df8c564e67f.png)
![Capture d’écran 2020-08-31 à 10 46 13](https://user-images.githubusercontent.com/222463/91704763-50a9d180-eb7c-11ea-9999-b7489a95b22b.png)
![Capture d’écran 2020-08-31 à 10 48 43](https://user-images.githubusercontent.com/222463/91704766-50a9d180-eb7c-11ea-8797-9ad24e24880c.png)
![Capture d’écran 2020-08-31 à 10 48 49](https://user-images.githubusercontent.com/222463/91704767-51426800-eb7c-11ea-8390-c944b353a99e.png)
![Capture d’écran 2020-08-31 à 10 48 55](https://user-images.githubusercontent.com/222463/91704768-51426800-eb7c-11ea-88a1-0699ce7cd281.png)
![Capture d’écran 2020-08-31 à 10 49 01](https://user-images.githubusercontent.com/222463/91704769-51dafe80-eb7c-11ea-8c68-b03243564d93.png)
![Capture d’écran 2020-08-31 à 10 49 07](https://user-images.githubusercontent.com/222463/91704770-51dafe80-eb7c-11ea-8e1e-f715323b19c8.png)
![Capture d’écran 2020-08-31 à 10 49 14](https://user-images.githubusercontent.com/222463/91704773-51dafe80-eb7c-11ea-8676-49822bb63020.png)
![Capture d’écran 2020-08-31 à 10 49 20](https://user-images.githubusercontent.com/222463/91704774-52739500-eb7c-11ea-97d7-3c37e7b2152a.png)
![Capture d’écran 2020-08-31 à 10 49 27](https://user-images.githubusercontent.com/222463/91704775-52739500-eb7c-11ea-95d3-40f257323daf.png)
![Capture d’écran 2020-08-31 à 10 49 33](https://user-images.githubusercontent.com/222463/91704777-52739500-eb7c-11ea-93f3-33a385385f27.png)
![Capture d’écran 2020-08-31 à 10 49 38](https://user-images.githubusercontent.com/222463/91704780-530c2b80-eb7c-11ea-8611-a32eebff17e6.png)
![Capture d’écran 2020-08-31 à 10 49 45](https://user-images.githubusercontent.com/222463/91704782-530c2b80-eb7c-11ea-898c-30862f818563.png)
![Capture d’écran 2020-08-31 à 10 49 50](https://user-images.githubusercontent.com/222463/91704783-53a4c200-eb7c-11ea-8a46-d0540ac320d0.png)
![Capture d’écran 2020-08-31 à 10 50 02](https://user-images.githubusercontent.com/222463/91704785-53a4c200-eb7c-11ea-8740-e9b25d30cf66.png)
![Capture d’écran 2020-08-31 à 10 50 12](https://user-images.githubusercontent.com/222463/91704787-543d5880-eb7c-11ea-948d-a3f96e2072f6.png)

## After

### Mean time: 54 seconds

Run | Valid | Errors | Real | User | Sys
----|-------|--------|------|------|----
1 | 34 | 17 | 0m51.077s | 0m41.154s | 0m2.916s
2 | 27 | 24 | 0m58.570s | 0m35.633s | 0m2.681s
3 | 33 | 18 | 0m53.012s | 0m43.293s | 0m2.954s

![Capture d’écran 2020-08-31 à 11 24 24](https://user-images.githubusercontent.com/222463/91705132-d9287200-eb7c-11ea-97f1-8cd03f0707d8.png)
![Capture d’écran 2020-08-31 à 11 24 36](https://user-images.githubusercontent.com/222463/91705135-d9c10880-eb7c-11ea-9239-0cb82b20cb70.png)
![Capture d’écran 2020-08-31 à 11 24 49](https://user-images.githubusercontent.com/222463/91705136-da599f00-eb7c-11ea-9677-8196b7761088.png)