@Ndpnt 2020-08-31
Also, can you please explain this: "Makes selector failure detection more accurate." because I did not see the difference.