@Ndpnt 2020-09-02
Can you take the opportunity of this PR to also remove the tosback-import script and its validator file?