@MattiSG 2020-09-03
- [ ] The validation command does not work with service name containing a space.
- [ ] When there is a schema validation error, it should log the service and the document which fails and display the error.
- [x] It should list which document has an error (not only the service).
- [ ] It should list all documents, which one fails and which one succeeds.
- [x] A mocha like summary is missing. I found it very useful.