@MattiSG 2020-09-03
I can't reproduce “The validation command does not work with service name containing a space”:

```
› npm run validate "Apple App Store"

> cgus@0.0.1 validate /Users/diplomatie/Documents/cgus
> node scripts/validation/validate.js "Apple App Store"

Validating 1 service declarations…
Apple App Store has valid Review Guidelines

1 services are valid
```