@MattiSG 2020-09-03
### Validating schemas

![Capture d’écran 2020-09-03 à 19 09 36](https://user-images.githubusercontent.com/222463/92145809-0f821d80-ee19-11ea-9188-c3ad985f66a8.png)
![Capture d’écran 2020-09-03 à 19 09 22](https://user-images.githubusercontent.com/222463/92145804-0d1fc380-ee19-11ea-8c66-c0dbdb617bd8.png)

### Validating services

![Capture d’écran 2020-09-03 à 19 10 50](https://user-images.githubusercontent.com/222463/92145970-4d7f4180-ee19-11ea-8f54-0756736ced5a.png)
![Capture d’écran 2020-09-03 à 19 11 00](https://user-images.githubusercontent.com/222463/92145974-4e17d800-ee19-11ea-8356-cc49d05a2772.png)
![Capture d’écran 2020-09-03 à 19 11 11](https://user-images.githubusercontent.com/222463/92145977-4eb06e80-ee19-11ea-8218-02283895ba36.png)
![Capture d’écran 2020-09-03 à 19 11 20](https://user-images.githubusercontent.com/222463/92145981-507a3200-ee19-11ea-9a17-9eb8f71cceb9.png)
![Capture d’écran 2020-09-03 à 19 11 29](https://user-images.githubusercontent.com/222463/92145986-51ab5f00-ee19-11ea-8535-0aa763278a7c.png)
![Capture d’écran 2020-09-03 à 19 11 34](https://user-images.githubusercontent.com/222463/92145991-53752280-ee19-11ea-82b5-af39b5b24d06.png)

### Timeouts

I'm starting to investigate why we have so many fetch timeouts with this new implementation.