@MattiSG 2020-09-03
I tried adding `await new Promise((resolve, reject) => setTimeout(resolve, Math.random() * 30000));` before each fetch, to no avail. I still get random failures, most of the time on Google services.

I thought that could be Google's DDoS protection, but if I run `npm run validate Google GoogleAdSense GoogleAnalytics GoogleAdMob GoogleAds GooglePlayStore` (with no random buffer time), they all pass.

I have no idea why we get random failures when we validate many services at once.