@MattiSG 2020-09-04
The parallelization issue is too big to investigate as of now. It will be part of a larger assessment of concurrency handling. Depending on the results of that assessment and potential remedies, this PR might be resurrected.

Meanwhile, this PR will be closed but the branch will be left in the repository. The main improvements demonstrated (relying on git rather than JS code to select changed services, selecting multiple services) will be backported to #110.